Solve React linking issues on pod install
Since the recent change of import statements from (import "RCTBridgeModule.h") to be like (import <React/RCTBridgeModule.h>), the pod install method was not properly linking the React library to the project. This addition of React as a dependency in the podspec fixed the issue of "<React/RCTBRidgeModule.h> not found" that I was getting.